refactored code so it's more readable
