+++ main.cpp
@@ -2,44 +2,18 @@
 #include <iostream>
 
 #include "constants.cpp"
-#include "food.cpp"
-#include "snake.cpp"
+#include "game.cpp"
 
 using namespace std;
-
-// Time passed since last game update
-double lastUpdateTime = 0;
-
-// Checks if a certain time interval since the last time it happened
-bool shouldTriggerUpdate(double updateInterval) { 
-    double currentTime = GetTime();
-    if (currentTime - lastUpdateTime >= updateInterval) {
-        lastUpdateTime = currentTime;
-        return true;
-    }
-    return false;
-}
-
-int lastDirectionKeyPressed = KEY_RIGHT;
-
-// Will change the last key pressed if it is a valid movement key
-void handleUserInput() {
-    int pressedKey = GetKeyPressed();
-    if (validMovementKeys.count(pressedKey)) {
-        lastDirectionKeyPressed = pressedKey;
-    }
-}
 
 int main() {
 
     // Init app
-    cout << "Starting the game..." << endl;
-    InitWindow(screenSize, screenSize, "test");
+    InitWindow(screenSize, screenSize, "snake game");
     SetTargetFPS(60);
 
     // Game setup
-    Food food = Food();
-    Snake snake = Snake();
+    Game game; 
 
     // Gameloop
     while (!WindowShouldClose()) {
@@ -47,16 +21,16 @@ int main() {
         
         // Computing
         
-        handleUserInput();
-        if(shouldTriggerUpdate(updateInterval)) {
-            snake.updateDirection(lastDirectionKeyPressed);
-            snake.update();
+        game.handleUserInput();
+        if(game.shouldTriggerUpdate(updateInterval)) {
+            game.snake.updateDirection(game.lastDirectionKeyPressed);
+            game.snake.update();
         }
 
         // Drawing
         ClearBackground(LGREEN);
-        food.draw();
-        snake.draw();
+        game.food.draw();
+        game.snake.draw();
 
         EndDrawing();
     }
